+++ .gitignore
@@ -2,3 +2,4 @@
 /.regolith
 /node_modules
 /package-lock.json
+/packs
