+++ package.json
@@ -1,15 +1,14 @@
 {
-  "name":"@conmaster/database",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "name": "con-database",
+  "main": "database.js",
+  "types": "database.d.ts",
   "type": "module",
   "files": [
-    "src/index.js",
-    "src/index.d.ts"
+    "database.js",
+    "database.d.ts"
   ],
   "dependencies": {
     "@minecraft/server": "^1.5.0-beta.1.20.20-preview.20",
-    "@minecraft/server-ui": "^1.2.0-beta.1.20.20-preview.20",
-    "nbt-serializer": "^1.0.1"
+    "@minecraft/server-ui": "^1.2.0-beta.1.20.20-preview.20"
   }
 }
