--- packs/BP/scripts/utils.js
@@ -1,1 +1,0 @@
-export * from "./utils/index.js";
