--- packs/BP/scripts/utils/index.js
@@ -1,2 +1,0 @@
-export * from "./nbt-serializer";
-export * from "./database";
